Move IP settings into shared module
